# backend-urbanfield/app/swagger.yml
swagger: "2.0"
info:
  title: Urban Fields API
  description: API dokumentasi untuk aplikasi booking lapangan olahraga Urban Fields
  version: "1.0.0"
basePath: /
schemes:
  - http
paths:
...
  /fields/{field_id}/booked_slots:
    get:
      tags: [Bookings]
      summary: Ambil daftar slot waktu yang sudah dibooking pada tanggal tertentu
      parameters:
        - in: path
          name: field_id
          required: true
          type: integer
        - in: query
          name: date
          required: true
          type: string
          description: "Format: YYYY-MM-DD"
      responses:
        200: { description: "List jam yang sudah terbooking, contoh: [\"10:00\", \"14:00\"]" }
        400: { description: Parameter date tidak diisi }
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Buat booking baru
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              user_id: { type: integer }
              field_id: { type: integer }
              date: { type: string, description: "Format YYYY-MM-DD" }
              time: { type: string, description: "Format HH:MM" }
              payment_method_id: { type: integer }
      responses:
        201: { description: Booking berhasil dibuat }
        409: { description: Slot waktu sudah dibooking orang lain }
    get:
      tags: [Bookings]
      summary: Ambil booking (semua, atau filter by user_id)
      parameters:
        - in: query
          name: user_id
          type: integer
          required: false
          description: Jika diisi, hanya kembalikan booking milik user ini
      responses:
        200: { description: List booking }
    put:
      tags: [Bookings]
      summary: Update booking
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id: { type: integer }
              user_id: { type: integer }
              field_id: { type: integer }
              date: { type: string }
              time: { type: string }
              payment_method_id: { type: integer }
      responses:
        200: { description: Booking berhasil diupdate }
        404: { description: Booking tidak ditemukan }
    delete:
      tags: [Bookings]
      summary: Batalkan/hapus booking (tidak bisa untuk booking yang tanggalnya sudah lewat)
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              id: { type: integer }
      responses:
        200: { description: Booking berhasil dibatalkan }
        400: { description: Booking sudah lewat tanggalnya, tidak bisa dibatalkan }
        404: { description: Booking tidak ditemukan }
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Masukkan token JWT dengan format: Bearer {token}"